calculations.py, config.py: Don't warn if close to value bounds
Example: Setting EV to 15, shutter speed to 1/125, and f-stop to 16, gives
exact ISO=98, resulting in a warning that ISO is out of bounds. Ideally,
since 98 should round to 100 but for example 84 should round to the next
lower standard ISO value, which is 80, the former should not yield a
warning while the latter should yield a warning. This is now implemented.